+++ packages/theme/src/color-ramps/stories/ramp-table.tsx
@@ -24,19 +24,16 @@ const RAMP_TOKENS_ORDER: { tokenName: keyof Ramp; abbr: string }[] = [
 	{ tokenName: 'bgFill2', abbr: 'BGF2' },
 	{ tokenName: 'bgFillInverted1', abbr: 'BGFI1' },
 	{ tokenName: 'bgFillInverted2', abbr: 'BGFI2' },
-	{ tokenName: 'bgFillDark', abbr: 'BGFD' },
 	{ tokenName: 'stroke1', abbr: 'ST1' },
 	{ tokenName: 'stroke2', abbr: 'ST2' },
 	{ tokenName: 'stroke3', abbr: 'ST3' },
 	{ tokenName: 'stroke4', abbr: 'ST4' },
-	{ tokenName: 'fgSurface1', abbr: 'FGS1' },
 	{ tokenName: 'fgSurface2', abbr: 'FGS2' },
 	{ tokenName: 'fgSurface3', abbr: 'FGS3' },
 	{ tokenName: 'fgSurface4', abbr: 'FGS4' },
 	{ tokenName: 'fgSurface5', abbr: 'FGS5' },
 	{ tokenName: 'fgFill', abbr: 'FGF' },
 	{ tokenName: 'fgFillInverted', abbr: 'FGFI' },
-	{ tokenName: 'fgFillDark', abbr: 'FGFD' },
 ];
 
 type RampTableProps = {
@@ -270,7 +267,6 @@ export const RampTable = forwardRef< HTMLDivElement, RampTableProps >(
 									'surface3',
 									'bgFill1',
 									'bgFillInverted1',
-									'bgFillDark',
 								].includes( tokenName ) ? (
 									<span
 										style={ {
@@ -285,13 +281,6 @@ export const RampTable = forwardRef< HTMLDivElement, RampTableProps >(
 									>
 										{ tokenName === 'surface3' ? (
 											<>
-												<span
-													style={ {
-														color: ramp.fgSurface1,
-													} }
-												>
-													Aa
-												</span>
 												<span
 													style={ {
 														color: ramp.fgSurface2,
@@ -335,15 +324,6 @@ export const RampTable = forwardRef< HTMLDivElement, RampTableProps >(
 											<span
 												style={ {
 													color: ramp.fgFillInverted,
-												} }
-											>
-												Aa
-											</span>
-										) : null }
-										{ tokenName === 'bgFillDark' ? (
-											<span
-												style={ {
-													color: ramp.fgFillDark,
 												} }
 											>
 												Aa